Fixed manage account when no account was signed in
