+++ client/src/Components/CustomLink.jsx
@@ -8,6 +8,7 @@ const CustomLink = styled(Link)(({ theme }) => ({
   position: "relative", // Position relative for pseudo-elements
   color: theme.palette.secondary.main, // Default color
   transition: "color 0.3s ease-out", // Smooth transition for color change
+  fontFamily: "Poppins",
 
   "&.active": {
     color: theme.palette.primary.main,
