+++ client/src/Pages/Dashboard/AnimalsPawrent.jsx
@@ -1,40 +1,31 @@
+// react and other functions
 import React, { useEffect, useState } from "react";
-import animalHero from "../assets/images/animals/animalHero.png";
-import {
-  Box,
-  Button,
-  Container,
-  Grid,
-  Typography,
-  Pagination,
-} from "@mui/material";
-import AnimalCard from "../Components/Card/AnimalCard";
-import Footer from "../Components/PageComponent/Footer";
-import AddIcon from "@mui/icons-material/Add";
 import axios from "axios";
-import { SortByButton } from "../Components/SortByButton";
-import { FilterOptions } from "../Components/FilterOptions";
-import ModalAddPet from "../Components/Modal/ModalAddPet";
-import { SearchInput } from "../Components/SearchInput";
-import PawrentCard from "../Components/Card/PawrentCard";
-import { SelectShelter } from "../Components/SelectShelter";
-import {
-  defaultUploadedImages,
-  filteredOptions,
-  defaultAnimalData,
-} from "../constants/animals";
-import { ascendingPets, descendingPets } from "../constants/sortLogic";
-import useAuth from "../hooks/useAuth";
 
-const AnimalsShelter = () => {
+// mui components
+import Container from "@mui/material/Container";
+import Grid from "@mui/material/Grid";
+import Typography from "@mui/material/Typography";
+
+// constant datas
+import { filteredOptions } from "../../constants/animals";
+
+// sorting logic
+import { ascendingPets, descendingPets } from "../../constants/sortLogic";
+
+// custom components
+import { FilterOptions } from "../../Components/FilterOptions";
+import { SelectShelter } from "../../Components/SelectShelter";
+import Footer from "../../Components/PageComponent/Footer";
+import PawrentCard from "../../Components/Card/PawrentCard";
+import { SortByButton } from "../../Components/SortByButton";
+import { SearchInput } from "../../Components/SearchInput";
+
+// icons and images
+import animalHero from "../assets/images/animals/animalHero.png";
+
+const AnimalsPawrent = () => {
   const [animals, setAnimals] = useState([]);
-  const { isLoggedIn, user, role } = useAuth();
-
-  console.log("Logged in: " + isLoggedIn);
-  console.log(user);
-  console.log(role);
-
-  const isShelter = role === "Rescue Shelter";
 
   console.table(animals);
 
@@ -51,7 +42,6 @@ const AnimalsShelter = () => {
       .then(function (response) {
         const allPets = response.data.allPets;
         setAnimals(allPets);
-        console.log(allPets);
       })
       .catch(function (error) {
         console.log(error);
@@ -69,12 +59,17 @@ const AnimalsShelter = () => {
 
       let filteredPets = response.data.filteredPets;
 
+      // Apply sorting logic to the filtered subset of animals
       if (sortBy === "Ascending") {
         filteredPets = filteredPets.sort((a, b) => {
+          // Sorting logic for ascending order
+          // For example, compare the name property of animals
           return a.name.localeCompare(b.name);
         });
       } else if (sortBy === "Descending") {
         filteredPets = filteredPets.sort((a, b) => {
+          // Sorting logic for descending order
+          // For example, compare the name property of animals
           return b.name.localeCompare(a.name);
         });
       }
@@ -103,9 +98,9 @@ const AnimalsShelter = () => {
     const selectedSortBy = e.target.value;
     setSortBy((prevSortBy) => {
       if (selectedSortBy === "Ascending") {
-        setAnimals(ascendingPets(animals));
+        setAnimals(ascendingPets);
       } else if (selectedSortBy === "Descending") {
-        setAnimals(descendingPets(animals));
+        setAnimals(descendingPets);
       }
       return selectedSortBy;
     });
@@ -124,6 +119,16 @@ const AnimalsShelter = () => {
     setSearchValue(event.target.value);
   };
 
+  // Select shelter
+  const [shelter, setShelter] = useState("");
+
+  const handleChangeShelter = (e) => {
+    const selectedShelter = e.target.value;
+
+    // TODO: Shelter change logic
+    setShelter(selectedShelter);
+  };
+
   // Pagination
   const itemsPerPage = 20;
   const [currentPage, setCurrentPage] = useState(1);
@@ -135,17 +140,6 @@ const AnimalsShelter = () => {
   const handlePageChange = (event, value) => {
     setCurrentPage(value);
   };
-
-  // Select shelter
-  const [shelter, setShelter] = useState("");
-
-  const handleChangeShelter = (e) => {
-    const selectedShelter = e.target.value;
-
-    // TODO: Shelter change logic
-    setShelter(selectedShelter);
-  };
-
   return (
     <>
       <div className="relative bg-gradient-to-bl from-amber-500 to-orange-600 h-[65vh] w-full flex justify-end items-center">
@@ -165,31 +159,8 @@ const AnimalsShelter = () => {
         ></div>
       </div>
       <Container maxWidth="lg" sx={{ my: "64px", position: "relative" }}>
-        <Container
-          maxWidth="md"
-          sx={{
-            textAlign: "center",
-            display: "flex",
-            justifyContent: "center",
-            flexDirection: "column",
-            alignItems: "center",
-            marginBottom: "32px",
-          }}
-        >
-          <Typography variant="h6" fontWeight={600} gutterBottom>
-            It’s My Pawssion!
-          </Typography>
-          <Typography variant="body2" fontWeight={300}>
-            Here is the list of animals in your shelter. You can add another
-            rescue, update some of their information, or remove adopted pets
-          </Typography>
-        </Container>
-
-        {/* Sort by, search, add / shelter component */}
+        <Content />
         <OptionSection
-          isShelter={isShelter}
-          isLoggedIn={isLoggedIn}
-          setAnimals={setAnimals}
           sortBy={sortBy}
           handleChangeSortBy={handleChangeSortBy}
           searchValue={searchValue}
@@ -197,71 +168,49 @@ const AnimalsShelter = () => {
           shelter={shelter}
           handleChangeShelter={handleChangeShelter}
         />
-
-        {/* Filter Component */}
         <FilterOptions
           filters={filteredOptions}
           handleFilterChange={handleFilterChange}
           handleApplyFilters={handleApplyFilters}
         />
-
-        {/* Display animal cards */}
         <DisplayAnimalCards
-          isShelter={isShelter}
-          isLoggedIn={isLoggedIn}
           currentAnimals={currentAnimals}
           setAnimals={setAnimals}
         />
-
-        {/* Page Component */}
-        <PageComponent
-          count={totalPages}
-          page={currentPage}
-          onChange={handlePageChange}
-        />
       </Container>
-
-      {/* Footer */}
       <Footer />
     </>
   );
 };
 
-export default AnimalsShelter;
+export default AnimalsPawrent;
 
-function AddAnimal({ setAnimals }) {
-  const [openAddModal, setOpenAddModal] = useState(false);
-
-  const handleModalClose = () => {
-    setOpenAddModal(false);
-  };
+const Content = () => {
   return (
-    <>
-      <Button
-        variant="outlined"
-        endIcon={<AddIcon />}
-        onClick={() => {
-          setOpenAddModal(true);
-        }}
-        sx={{ textTransform: "none", padding: "6px 34px", width: "215px" }}
-      >
-        Add New Pet
-      </Button>
-      <ModalAddPet
-        open={openAddModal}
-        onClose={handleModalClose}
-        defaultAnimal={defaultAnimalData}
-        defaultImage={defaultUploadedImages}
-        setAnimals={setAnimals}
-      />
-    </>
+    <Container
+      maxWidth="md"
+      sx={{
+        textAlign: "center",
+        display: "flex",
+        justifyContent: "center",
+        flexDirection: "column",
+        alignItems: "center",
+        marginBottom: "32px",
+      }}
+    >
+      <Typography variant="h6" fontWeight={600} gutterBottom>
+        It’s My Pawssion!
+      </Typography>
+      <Typography variant="body2" fontWeight={300} width="600px">
+        Here is the list of animals in your shelter. You can add another rescue,
+        update some of their information, or remove adopted pets
+      </Typography>
+    </Container>
   );
-}
+};
 
+// TODO: Select shelter
 const OptionSection = ({
-  isShelter,
-  isLoggedIn,
-  setAnimals,
   sortBy,
   handleChangeSortBy,
   searchValue,
@@ -294,22 +243,13 @@ const OptionSection = ({
         <SearchInput value={searchValue} onChange={handleSearchInputChange} />
       </Grid>
       <Grid item order={{ xs: 2, sm: 3, md: 3, lg: 3 }}>
-        {isShelter && isLoggedIn ? (
-          <AddAnimal setAnimals={setAnimals} />
-        ) : (
-          <SelectShelter value={shelter} onChange={handleChangeShelter} />
-        )}
+        <SelectShelter value={shelter} onChange={handleChangeShelter} />
       </Grid>
     </Grid>
   );
 };
 
-const DisplayAnimalCards = ({
-  isShelter,
-  isLoggedIn,
-  currentAnimals,
-  setAnimals,
-}) => {
+const DisplayAnimalCards = ({ currentAnimals, setAnimals }) => {
   return (
     <Grid container columnSpacing={3} rowSpacing={4} mt={4} alignItems="center">
       {currentAnimals.length === 0 ? (
@@ -333,50 +273,15 @@ const DisplayAnimalCards = ({
             display="flex"
             justifyContent="center"
           >
-            {isShelter && isLoggedIn ? (
-              <AnimalCard
-                animals={animal}
-                height="auto"
-                width="257px"
-                setAnimals={setAnimals}
-              />
-            ) : (
-              <PawrentCard
-                animals={animal}
-                height="auto"
-                width="257px"
-                setAnimals={setAnimals}
-              />
-            )}
+            <PawrentCard
+              animals={animal}
+              height="auto"
+              width="257px"
+              setAnimals={setAnimals}
+            />
           </Grid>
         ))
       )}
     </Grid>
-  );
-};
-
-const PageComponent = ({ count, page, onChange }) => {
-  return (
-    <Box
-      sx={{
-        width: "100%",
-        justifyContent: "center",
-        display: "flex",
-        mt: 2,
-      }}
-    >
-      <Pagination
-        variant="rounded"
-        count={count}
-        page={page}
-        onChange={onChange}
-        size="large"
-        siblingCount={1}
-        boundaryCount={1}
-        showFirstButton
-        showLastButton
-        sx={{ mt: 2, justifyContent: "center" }}
-      />
-    </Box>
   );
 };
