+++ client/src/renderNavbar.jsx
@@ -17,7 +17,8 @@ export const renderNavbar = () => {
     "/donate",
     "/contact",
     "/animalsPawrent",
-    "/manage",
+    "/manage/pawrent",
+    "/manage/shelter",
     "/test",
     "/reqshelter",
     "/reqpawrent",
